+++ app/Http/Controllers/StatusController.php
@@ -102,8 +102,17 @@ class StatusController extends Controller
     public function destroy(Status $status)
     {
         //Elimina el registro de la base de datos
-        $status->delete();
+        /*$status->delete();
         session()->flash('statusKey', 'Status was deleted!');
-        return to_route('status.index');
+        return to_route('status.index');*/
+
+         if ($status->packages->count() > 0) {
+             return to_route('status.index');
+         } else {
+             //Elimina el registro de la base de datos
+             $status->delete();
+             session()->flash('statusKey', 'Status was deleted!');
+             return to_route('status.index');
+         }
     }
 }
